Align disk images on 4k sectors
The upstream openwrt images are not aligned on 4k sector boundaries. This leads
to qemu throwing errors about lacking the resize permission because the image
is opened with O_DIRECT which expects the image to be aligned and will try to
align it automatically if it doesn't. But that automatic alignment requires the
RESIZE permission in qemu.

During docker image builds we resize the upstream image by a couple MB and
specify 'M' to dd instead of 'MB'. The difference is documented in the dd man
page but boils down to the difference of 1000*1000 (MB) versus 1024*1024 (M).
By using 'M' we ensure an image aligned on 4096 bytes

Additionally this commit modifies the IMAGE_SIZE_ON_INIT code so it also uses
'M' and generates 4k aligned images as well.

Testing: Built a docker image with my changes and ran it